Added service classes for service layer.
Refactored pmid object type to long.

Added controller classes for RESTful API.

+++ src/main/java/reciter/algorithm/cluster/Clusterer.java
@@ -22,5 +22,5 @@ public interface Clusterer {
 	
 	List<ReCiterArticle> getReCiterArticles();
 	
-	Map<Integer, ReCiterCluster> getClusters();
+	Map<Long, ReCiterCluster> getClusters();
 }
